change discovery logic
