+++ aiohue/__init__.py
@@ -3,4 +3,4 @@
 from .v1 import HueBridgeV1  # noqa
 from .v2 import HueBridgeV2  # noqa
 from .errors import *  # noqa
-from .util import create_app_key, is_v2_bridge  # noqa
+from .util import create_app_key  # noqa
